requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-jazzmin==2.6.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==2.0.2
wsproto==1.2.0
